Make patchChildren accept a function
