+++ packages/fuselage/src/helpers/flattenChildren.js
@@ -1,0 +1,1 @@
+export { default as flattenChildren } from 'react-keyed-flatten-children';
