+++ packages/fuselage/src/helpers/memoize.js
@@ -1,14 +1,11 @@
 const store = new WeakMap();
 
-const getKey = (arg) => arg;
-
 export const memoize = (fn) => {
   const cache = new Map();
-  const memoized = function(arg) {
-    const key = getKey(arg);
 
-    if (cache.has(key)) {
-      return cache.get(key);
+  const memoized = function(arg) {
+    if (cache.has(arg)) {
+      return cache.get(arg);
     }
 
     const result = fn.call(this, arg);
@@ -17,7 +14,9 @@ export const memoize = (fn) => {
 
     return result;
   };
+
   store.set(memoized, cache);
+
   return memoized;
 };
